Move Datasets REST API to separated module
