+++ backend/medtagger/api/datasets/__init__.py
@@ -1,0 +1,1 @@
+"""Module responsible for endpoints returning and creating Datasets."""
