+++ frontend/src/app/utils/ApiUrl.ts
@@ -2,7 +2,7 @@ export enum API_URL {
     LOGIN = '/auth/sign-in',
     REGISTER = '/auth/register',
     USER_INFO = '/users/info',
-    DATASETS = '/scans/datasets',
+    DATASETS = '/datasets',
     LABEL = '/label',
     RANDOM_SCAN = '/scans/random',
     SCANS = '/scans/',
